projects/sidekick-contracts/pyproject.toml
[tool.poetry]
name = "sidekick-contracts"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["User <user@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
algokit-utils = "^4.0.0"
python-dotenv = "^1.0.0"
algorand-python = "^2.0.0"
algorand-python-testing = "~0"
beaker-pyteal = "^1.0.0"
pyteal = "^0.25.0"

[tool.poetry.group.dev.dependencies]
algokit-client-generator = "^2.1.0"
puyapy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

